# docs/OPENAPI.yaml
...
openapi: 3.1.0
info:
  title: Oblivion API
  description: 'Backend API for controlled data erasure, recovery, residual analysis, assurance and certificate
    verification.


    This document is GENERATED by scripts/gen_openapi.py from the FastAPI application itself. Do not hand-edit it:
    edit the routes and schemas, then regenerate.


    Authentication uses HTTP Bearer with opaque server-side session tokens (Argon2id password hashing, SHA-256 token
    digest stored server-side). These are NOT JWTs. There is no refresh-token endpoint and no cookie based session;
    the bearer token is the only credential transport.'
  version: 0.1.0
paths:
  /api/auth/login:
    post:
      tags:
      - auth
      summary: Login
      description: Authenticates a user and issues an opaque server-side session token.
      operationId: login_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security: []
  /api/auth/logout:
    post:
      tags:
      - auth
      summary: Logout
      description: Revokes the active session token.
      operationId: logout_api_auth_logout_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security: []
  /api/auth/me:
    get:
      tags:
      - auth
      summary: Get Current User Profile
      description: Returns the profile and granted permissions of the authenticated user.
      operationId: get_current_user_profile_api_auth_me_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/audit/events:
    get:
      tags:
      - audit
      summary: List Audit Events
      description: 'Read audit records, oldest first (requires ``audit.view``).


        Unrecognised filter values are rejected with 422 rather than ignored. A

        filter that is silently dropped returns *more* than was asked for while

        looking like it returned exactly what was asked for - in an audit context,

        a way to miss the record you were told to find.'
      operationId: list_audit_events_api_audit_events_get
      parameters:
      - name: operation_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Operation Id
      - name: actor_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Actor Id
      - name: event_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Event Type
      - name: outcome
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Outcome
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Since
      - name: until
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Until
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventPageOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/audit/verify:
    post:
      tags:
      - audit
      summary: Verify Audit Chain Endpoint
      description: 'Verify the persisted audit chain (requires ``audit.verify``).


        Records are loaded from storage and rehashed. The caller supplies no

        digests, no records and no expected outcome - there is no field in this

        request in which to assert that the log is intact, which is what keeps the

        answer the server''s rather than the caller''s.'
      operationId: verify_audit_chain_endpoint_api_audit_verify_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/AuditVerifyRequest'
              - type: 'null'
              title: Request
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditChainVerificationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AuditChainVerificationOut:
      properties:
        status:
          type: string
          title: Status
        checked:
          type: integer
          title: Checked
        reason:
          type: string
          title: Reason
        links:
          items:
            $ref: '#/components/schemas/AuditLinkOut'
          type: array
          title: Links
        first_invalid_audit_id:
          anyOf:
          - type: string
          - type: 'null'
          title: First Invalid Audit Id
        records_predating_chain:
          type: integer
          title: Records Predating Chain
          default: 0
        proves:
          items:
            type: string
          type: array
          title: Proves
        does_not_prove:
          items:
            type: string
          type: array
          title: Does Not Prove
        scope_note:
          type: string
          title: Scope Note
          default: This result describes the integrity of the audit log only. It is not a statement about evidence
            integrity, erasure success, or certificate trust, each of which is verified separately against its own
            artefact.
      type: object
      required:
      - status
      - checked
      - reason
      - links
      - proves
      - does_not_prove
      title: AuditChainVerificationOut
      description: The result of verifying the persisted audit chain.
    AuditEventOut:
      properties:
        audit_id:
          type: string
          title: Audit Id
        sequence:
          type: integer
          title: Sequence
        event_type:
          type: string
          title: Event Type
        outcome:
          type: string
          title: Outcome
        actor_id:
          type: string
          title: Actor Id
        actor_role:
          type: string
          title: Actor Role
        actor_source:
          type: string
          title: Actor Source
        operation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Operation Id
        target_identity:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Identity
        evidence_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Evidence Id
        certificate_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Certificate Id
        summary:
          type: string
          title: Summary
          default: ''
        safe_metadata:
          additionalProperties: true
          type: object
          title: Safe Metadata
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        digest:
          type: string
          title: Digest
        previous_audit_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Previous Audit Id
        previous_audit_digest:
          anyOf:
          - type: string
          - type: 'null'
          title: Previous Audit Digest
      type: object
      required:
      - audit_id
      - sequence
      - event_type
      - outcome
      - actor_id
      - actor_role
      - actor_source
      - occurred_at
      - digest
      title: AuditEventOut
      description: One audit record as published.
    AuditEventPageOut:
      properties:
        events:
          items:
            $ref: '#/components/schemas/AuditEventOut'
          type: array
          title: Events
        returned:
          type: integer
          title: Returned
        total_records:
          type: integer
          title: Total Records
        records_predating_chain:
          type: integer
          title: Records Predating Chain
          default: 0
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - events
      - returned
      - total_records
      - limit
      - offset
      title: AuditEventPageOut
      description: 'A page of audit records, oldest first.


        ``total_records`` is the height of the whole log, not the size of this

        page, so a reader can tell a page is a window without inferring it from the

        page happening to be full.'
    AuditLinkOut:
      properties:
        audit_id:
          type: string
          title: Audit Id
        sequence:
          type: integer
          title: Sequence
        status:
          type: string
          title: Status
        detail:
          type: string
          title: Detail
      type: object
      required:
      - audit_id
      - sequence
      - status
      - detail
      title: AuditLinkOut
    AuditVerifyRequest:
      properties:
        operation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Operation Id
      type: object
      title: AuditVerifyRequest
      description: 'Optional narrowing for a verification request.


        Supplying ``operation_id`` narrows which records are examined. It cannot

        make a window look complete: a subset that does not begin at the genesis

        record is reported ``UNVERIFIABLE``, because the chain spans every event in

        the system and a per-operation slice necessarily has history before it.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        username:
          type: string
          maxLength: 128
          minLength: 1
          title: Username
        password:
          type: string
          maxLength: 256
          minLength: 1
          title: Password
      type: object
      required:
      - username
      - password
      title: LoginRequest
    LoginResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_at:
          type: string
          format: date-time
          title: Expires At
        user:
          $ref: '#/components/schemas/UserOut'
      type: object
      required:
      - access_token
      - expires_at
      - user
      title: LoginResponse
    LogoutResponse:
      properties:
        status:
          type: string
          title: Status
          default: COMPLETED
        message:
          type: string
          title: Message
          default: Session successfully terminated
      type: object
      title: LogoutResponse
...
    UserOut:
      properties:
        id:
          type: string
          title: Id
        username:
          type: string
          title: Username
        roles:
          items:
            type: string
          type: array
          title: Roles
        permissions:
          items:
            type: string
          type: array
          title: Permissions
        disabled:
          type: boolean
          title: Disabled
          default: false
        created_at:
          type: string
          format: date-time
          title: Created At
        last_authenticated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Authenticated At
      type: object
      required:
      - id
      - username
      - created_at
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OpaqueSessionToken
      description: 'Opaque session token issued by POST /api/auth/login. Send as ''Authorization: Bearer <token>''.
        The server stores only the SHA-256 digest and revokes it on logout.'
security:
- bearerAuth: []